Changed X and Y coordinates for Actors to floats. This will allow future support for being within a subsection of a tile. For example tile 15.6

+++ DivineRightConcept/DivineRightConcept/GameObjects/Actor.cs
@@ -11,13 +11,13 @@ namespace DivineRightConcept.GameObjects
     //An Actor class that should ideally be inherited for more precise functionality
     public class Actor
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
 
         //the current representation of the Actor which should be appriopiatly updated depending on its state
         public Texture2D Representation { get; set; }
 
-        public Actor(int X, int Y)
+        public Actor(float X, float Y)
         {
             this.X = X;
             this.Y = Y;
